refactor(settings): split mouse preview types by visualizer
